ToDictionary: Move ConfigToDictionaryConverter to separate file
Extract the converter to its own file to separate functionality from testing behavior
