+++ clean-code-katas/clean-code-katas.sln
@@ -21,6 +21,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{D06B579B
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "todictionary.tests", "todictionary.tests\todictionary.tests.csproj", "{B4F85FD8-A834-4A40-B55F-4887D0AC8F94}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "todictionary", "todictionary\todictionary.csproj", "{8BEDB266-9DCA-4BE6-B3C3-152095E71D69}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -31,6 +33,10 @@ Global
 		{B4F85FD8-A834-4A40-B55F-4887D0AC8F94}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B4F85FD8-A834-4A40-B55F-4887D0AC8F94}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B4F85FD8-A834-4A40-B55F-4887D0AC8F94}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8BEDB266-9DCA-4BE6-B3C3-152095E71D69}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8BEDB266-9DCA-4BE6-B3C3-152095E71D69}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8BEDB266-9DCA-4BE6-B3C3-152095E71D69}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8BEDB266-9DCA-4BE6-B3C3-152095E71D69}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -38,6 +44,7 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{D06B579B-4270-40D4-AE72-481828C5C9DC} = {A6B4F3FA-686E-44FC-B765-E35A7010999D}
 		{B4F85FD8-A834-4A40-B55F-4887D0AC8F94} = {D06B579B-4270-40D4-AE72-481828C5C9DC}
+		{8BEDB266-9DCA-4BE6-B3C3-152095E71D69} = {A6B4F3FA-686E-44FC-B765-E35A7010999D}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {67C6EEA4-00F5-49D9-B1C1-68D845BC7AFD}
