+++ clean-code-katas/todictionary.tests/todictionary.tests.csproj
@@ -68,6 +68,12 @@
   <ItemGroup>
     <Analyzer Include="..\packages\xunit.analyzers.0.10.0\analyzers\dotnet\cs\xunit.analyzers.dll" />
   </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\todictionary\todictionary.csproj">
+      <Project>{8BEDB266-9DCA-4BE6-B3C3-152095E71D69}</Project>
+      <Name>todictionary</Name>
+    </ProjectReference>
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
     <PropertyGroup>
